IMX-2071 BUG26867 CL277413 [NXP][android-Vulkan] Vulkan appliction crash on imx8qm/8qxp. 1) When the DU is enabled, we need to udpate the DU information when deleting a instruction. 2) Set the correct swizzle when merging ADD/LOAD instructions.
When the DU is enabled, we need to udpate the DU information when removing a instruction.
Move the STORE_ATTR of perpatch address instruction from the end of the shader to the beginning of the shader, so we may shorten the live range.
Fix a issue in VIR_Operand_NegateOperand: use the component type to do the negate/abs. Use case: dEQP-VK.tessellation.winding.default_domain.glsl_triangles_ccw.Don't set a vector type for a immediate operand.
Support UploadUBO for the new CG.
Fix some issues for big endian: remove non-32bit variables in ConstantValueUnion.
support UploadUBO for all graphic shaders, including TCS/TES/GS/CS.

+++ driver/compiler/libVSC/old_impl/optimizer/gc_vsc_old_optimizer.c
@@ -7410,7 +7410,7 @@ _GetLTCValue(
     tempValue.sourceInfo = 0;
     for (i = 0; i < MAX_LTC_COMPONENTS; i++)
     {
-        tempValue.v[i].u64 = 0;
+        tempValue.v[i].u32 = 0;
     }
 
     /* Get the uniform data. */
